requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
idna==3.4
iniconfig==2.0.0
isort==5.12.0
numpy==1.25.0
packaging==23.1
pandas==2.0.3
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.26.0
rest-condition==1.0.3
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
